Fix argument data type in LPF library (#317)

+++ src/shared/Filters/Inc/SharedFilters.h
@@ -50,12 +50,13 @@
  * @param input Pointer to an array of input samples
  * @param output Pointer to an array of output samples
  * @param sampling_time Sampling time interval for input and output samples
+ *                      in seconds
  * @param rc RC time constant
  */
 void SharedFilters_LowPassFilter(
     float32_t *input,
     float32_t *output,
-    uint32_t   sampling_time,
-    uint32_t   rc);
+    float32_t  sampling_time,
+    float32_t  rc);
 
 #endif /* SHARED_FILTERS_H */
